run integration tests only once
make it deploy only if all tests pass
